Move User routes to `route`
