+++ src/test/scala/com/hhandoko/realworld/route/UserRoutesSpec.scala
@@ -1,4 +1,4 @@
-package com.hhandoko.realworld.user
+package com.hhandoko.realworld.route
 
 import scala.concurrent.ExecutionContext
 
@@ -10,18 +10,20 @@ import org.specs2.matcher.MatchResult
 
 import com.hhandoko.realworld.auth.RequestAuthenticator
 import com.hhandoko.realworld.core.{JwtToken, User, Username}
+import com.hhandoko.realworld.route
+import com.hhandoko.realworld.user.UserService
 
 class UserRoutesSpec extends Specification { def is = s2"""
 
   User routes
     when logged in
-      should return 200 OK status             $uriReturns200
-      should return user info                 $uriReturnsUserInfo
+      should return 200 OK status             $currentUserReturns200
+      should return user info                 $currentUserReturnsUserInfo
     when not logged in
-      should return 401 Unauthorized status   $uriReturns401
+      should return 401 Unauthorized status   $unauthorisedReturns401
   """
   private[this] val nonExpiringToken = JwtToken("eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJyZWFsd29ybGQiLCJ1c2VybmFtZSI6ImZhbW91cyJ9.c3ghryIJayjtL3wL4j2KSEeLBXUd5U8ALbdSQBau2Qg")
-  private[this] val invalidToken = "invalid.jwt.token"
+  private[this] val invalidToken     = "invalid.jwt.token"
 
   private[this] val retCurrentUser: Response[IO] = {
     implicit val cs: ContextShift[IO] = IO.contextShift(ExecutionContext.global)
@@ -46,18 +48,18 @@ class UserRoutesSpec extends Specification { def is = s2"""
       headers = Headers.of(Header("Authorization", s"Token ${invalidToken}"))
     )
 
-    UserRoutes[IO](new RequestAuthenticator[IO], FakeUserService)
+    route.UserRoutes[IO](new RequestAuthenticator[IO], FakeUserService)
       .orNotFound(getCurrentUser)
       .unsafeRunSync()
   }
 
-  private[this] def uriReturns200: MatchResult[Status] =
+  private[this] def currentUserReturns200: MatchResult[Status] =
     retCurrentUser.status must beEqualTo(Status.Ok)
 
-  private[this] def uriReturnsUserInfo: MatchResult[String] =
+  private[this] def currentUserReturnsUserInfo: MatchResult[String] =
     retCurrentUser.as[String].unsafeRunSync() must beEqualTo(s"""{"user":{"email":"famous@test.com","token":"${nonExpiringToken.value}","username":"famous","bio":null,"image":null}}""")
 
-  private[this] def uriReturns401: MatchResult[Status] =
+  private[this] def unauthorisedReturns401: MatchResult[Status] =
     retUnauthorized.status must beEqualTo(Status.Unauthorized)
 
   object FakeUserService extends UserService[IO] {
@@ -65,5 +67,4 @@ class UserRoutesSpec extends Specification { def is = s2"""
       Some(User(username, None, None, s"${username.value}@test.com", nonExpiringToken))
     }
   }
-
 }
